server/requirements.txt
fastapi
uvicorn
python-jose
bcrypt
sqlalchemy
pydantic
python-dotenv
requests
